refactor: move npm resolve related files to one dir
