+++ src/resolve/npm/loadPackageMeta.ts
