+++ tsconfig.json
@@ -75,10 +75,10 @@
     "src/resolve/git.ts",
     "src/resolve/index.ts",
     "src/resolve/local.ts",
-    "src/resolve/npm.ts",
+    "src/resolve/npm/createNpmPkgId.ts",
+    "src/resolve/npm/index.ts",
+    "src/resolve/npm/loadPackageMeta.ts",
     "src/resolve/tarball.ts",
-    "src/resolve/utils/createNpmPkgId.ts",
-    "src/resolve/utils/loadPackageMeta.ts",
     "src/runScript.ts",
     "src/save.ts",
     "src/types.ts",
